fix failed to build a image while deleting it at the same time



See merge request docker/docker!405

+++ image/store.go
@@ -11,11 +11,17 @@ import (
 	"github.com/docker/docker/layer"
 )
 
+var (
+	ErrReferenceCountNotZero = errors.New("can not remove image: reference is not zero")
+)
+
 // Store is an interface for creating and accessing images
 type Store interface {
 	Create(config []byte) (ID, error)
 	Get(id ID) (*Image, error)
 	Delete(id ID) ([]layer.Metadata, error)
+	HoldOn(id ID) error
+	HoldOff(id ID, remove bool) error
 	Search(partialID string) (ID, error)
 	SetParent(id ID, parent ID) error
 	GetParent(id ID) (ID, error)
@@ -34,8 +40,9 @@ type LayerGetReleaser interface {
 }
 
 type imageMeta struct {
-	layer    layer.Layer
-	children map[ID]struct{}
+	layer          layer.Layer
+	children       map[ID]struct{}
+	referenceCount int
 }
 
 type store struct {
@@ -94,8 +101,9 @@ func (is *store) restore() error {
 		}
 
 		imageMeta := &imageMeta{
-			layer:    l,
-			children: make(map[ID]struct{}),
+			layer:          l,
+			children:       make(map[ID]struct{}),
+			referenceCount: 1,
 		}
 
 		is.images[ID(id)] = imageMeta
@@ -166,8 +174,9 @@ func (is *store) Create(config []byte) (ID, error) {
 	}
 
 	imageMeta := &imageMeta{
-		layer:    l,
-		children: make(map[ID]struct{}),
+		layer:          l,
+		children:       make(map[ID]struct{}),
+		referenceCount: 1,
 	}
 
 	is.images[imageID] = imageMeta
@@ -218,10 +227,18 @@ func (is *store) Get(id ID) (*Image, error) {
 func (is *store) Delete(id ID) ([]layer.Metadata, error) {
 	is.Lock()
 	defer is.Unlock()
+	return is.delete(id)
+}
+func (is *store) delete(id ID) ([]layer.Metadata, error) {
 
 	imageMeta := is.images[id]
 	if imageMeta == nil {
 		return nil, fmt.Errorf("unrecognized image ID %s", id.String())
+	}
+
+	imageMeta.referenceCount--
+	if imageMeta.referenceCount > 0 {
+		return nil, ErrReferenceCountNotZero
 	}
 	for id := range imageMeta.children {
 		is.fs.DeleteMetadata(id, "parent")
@@ -242,6 +259,74 @@ func (is *store) Delete(id ID) ([]layer.Metadata, error) {
 	}
 
 	return nil, nil
+}
+
+// HoldOn will increase the reference count for image
+func (is *store) HoldOn(id ID) error {
+	is.Lock()
+	defer is.Unlock()
+
+	var (
+		iter ID = id
+		err  error
+	)
+	if imageMeta := is.images[id]; imageMeta == nil {
+		return fmt.Errorf("unrecognized image ID %s", id.String())
+	}
+	for {
+		imageMeta := is.images[iter]
+		if imageMeta == nil {
+			break
+		}
+		imageMeta.referenceCount++
+		if iter, err = is.GetParent(iter); err != nil {
+			break
+		}
+
+	}
+	return nil
+}
+
+// HoldOff will decrease the reference count for image
+func (is *store) HoldOff(id ID, remove bool) error {
+	is.Lock()
+	defer is.Unlock()
+
+	var (
+		err        error
+		iter       ID   = id
+		needRemove bool = false
+	)
+	imageMeta := is.images[id]
+	if imageMeta == nil {
+		return fmt.Errorf("unrecognized image ID %s", id.String())
+	}
+	if remove && imageMeta.referenceCount == 1 {
+		needRemove = true
+	}
+	for {
+		imageMeta := is.images[iter]
+		if imageMeta == nil {
+			break
+		}
+		imageMeta.referenceCount--
+		if !remove && imageMeta.referenceCount == 0 {
+			// If need to keep this image, and the reference count is zero
+			// here need to reset it to 1. 1 stands for could be removed if a delete operation comes.
+			imageMeta.referenceCount = 1
+		}
+
+		if iter, err = is.GetParent(iter); err != nil {
+			break
+		}
+	}
+	if needRemove {
+		if _, err := is.delete(id); err != nil {
+			logrus.Errorf("Deleting image %s error: %v", id.String(), err)
+			return err
+		}
+	}
+	return nil
 }
 
 func (is *store) SetParent(id, parent ID) error {
